+++ src/main/java/org/example/multimedia_file_security/threadLocal/UserThreadLocal.java
@@ -1,13 +1,13 @@
 package org.example.multimedia_file_security.threadLocal;
 
 public class UserThreadLocal {
-    public static ThreadLocal<Integer> userThreadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Long> userThreadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Integer id) {
+    public static void setCurrentId(Long id) {
         userThreadLocal.set(id);
     }
 
-    public static Integer getCurrentId() {
+    public static Long getCurrentId() {
         return userThreadLocal.get();
     }
